+++ includes/autoload.php
@@ -3,7 +3,7 @@
  |  Snicker     The first native FlatFile Comment Plugin 4 Bludit
  |  @file       ./includes/autoload.php
  |  @author     SamBrishes <sam@pytes.net>
- |  @version    0.1.0 [0.1.0] - Alpha
+ |  @version    0.1.1 [0.1.0] - Alpha
  |
  |  @website    https://github.com/pytesNET/snicker
  |  @license    X11 / MIT License
@@ -11,8 +11,11 @@
  */
 
     spl_autoload_register(function($class){
-        if(strpos($class, "Gregwar") === 0 || strpos($class, "Identicon") === 0  || strpos($class, "PIT") === 0){
-			$path = dirname(__FILE__) . DIRECTORY_SEPARATOR;
+        foreach(array("Gregwar", "Identicon",  "PIT", "OWASP") AS $allowed){
+            if(strpos($class, $allowed) !== 0){
+                continue;
+            }
+            $path = dirname(__FILE__) . DIRECTORY_SEPARATOR;
 			$class = str_replace("\\", DIRECTORY_SEPARATOR, $class);
             require_once $class . ".php";
         }