+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+Version 0.1.1 - Alpha
+---------------------
+-   Add: The Persian Translation, many thanks to [abdulhalim](https://github.com/abdulhalim).
+-   Add: The GD-less PureCaptcha library, written by Abbas Naderi, which doesn't use PHP GD.
+         Thanks to [#8](https://github.com/pytesNET/snicker/issues/8)
+-   Add: Reload function to the Captcha Image (Click on the image).
+         Thanks to [#5](https://github.com/pytesNET/snicker/issues/5).
+-   Add: A JS-Snippet which Enables/Disables the Gravatar Select Field.
+         Thanks to [#9](https://github.com/pytesNET/snicker/issues/9).
+-   Add: The Captcha and Terms note aren't shown for logged in users.
+         Thanks to [#7](https://github.com/pytesNET/snicker/issues/7).
+-   Bugfix: The `comments_per_page` option couldn't be set on 0 to disable the limit!
+            Thanks to [#9](https://github.com/pytesNET/snicker/issues/9).
+-   Bugfix: Reload the Captcha, if it was wrong.
+            Thansk to [#5](https://github.com/pytesNET/snicker/issues/5).
+
 Version 0.1.0 - Alpha
 ---------------------
 -   Initial Release owo/